Include only binary packages in apt repo
Some of our packages have dynamically generated orig.tar.gz, so it will
differ each time (at least because of included timestamps), even if
every file inside is the same. This is unacceptable for reprepro.
Fixing this properly would require some rework in package build process,
so for now simply do not include source package in the repo. Sources are
always available in git repos.