Set default text/plain handler to GUI application
Double-clicking on a text file in GUI file manager shouldn't open text
editor if there is GUI one installed.
If not specified explicitly, the default depends on installation order,
which may be not deterministic. In practice, vim was being set.

Fix this by setting default to gedit (if present) and in later order,
any GUI editor (the one without Termina=true).

QubesOS/qubes-issues#4195