Use host's reprepro if available
In preparation for pbuilder usage, try to not strictly depend on custom
chroot environment for update-repo-* targets. reprepro is packaged for
major distributions already, so using version installed on the host is
an option. If it isn't there, fallback to the old approach, but extract
complexity into a helper script and keep Makefile.debian simpler.